requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0 # Updated from scipy>=1.10.0 on 16 October 2025

# Large-scale data processing
dask[complete]>=2023.1.0
pyarrow>=11.0.0
polars>=0.20.0

# Geospatial and Arctic-specific
cartopy>=0.21.0
geopandas>=0.14.0
shapely>=2.0.0
rasterio>=1.3.0
pyproj>=3.6.0 # Updated from pyproj>=3.5.0 on 16 October 2025
gdal>=3.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
cmocean>=3.0.0
cmasher>=1.6.0
contextily>=1.3.0
folium>=0.14.0

# Machine learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0

# Data formats
tables>=3.8.0
h5py>=3.8.0
netCDF4>=1.6.0
xarray>=2023.1.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
tabulate>=0.9.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0


# ==========================================
# Part II: GeoCryoAI Teacher Forcing
# ==========================================

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Model Explainability
lime>=0.2.0
shap>=0.41.0

# Hyperparameter Optimization
scikit-optimize>=0.9.0
optuna>=3.1.0

# Configuration Management
pyyaml>=6.0
hydra-core>=1.3.0

# Experiment Tracking
tensorboard>=2.12.0